Create partial class for SingleOrError and FirstOrError
