--- src/Lemonad.ErrorHandling/EnumerableExtensions/SingleOrErrorCase.cs
@@ -1,9 +1,0 @@
-namespace Lemonad.ErrorHandling.EnumerableExtensions {
-    /// <summary>
-    /// Represents the cases in the SingleOrError extension method for <see cref="System.Linq.Enumerable"/>.
-    /// </summary>
-    public enum SingleOrErrorCase {
-        NoElement,
-        ManyElements
-    }
-}
