+++ src/Lemonad.ErrorHandling/EnumerableFunctions/README.md
